♻️ Give Starlette its own integration module
